+++ cienv/cienv.go
@@ -41,6 +41,7 @@ func GetBuildInfo() (prInfo *BuildInfo, isPR bool, err error) {
 	owner, repo := getOwnerAndRepoFromSlug([]string{
 		"TRAVIS_REPO_SLUG",
 		"DRONE_REPO", // drone<=0.4
+		"BITBUCKET_REPO_FULL_NAME",
 	})
 	if owner == "" {
 		owner = getOneEnvValue([]string{
@@ -74,6 +75,7 @@ func GetBuildInfo() (prInfo *BuildInfo, isPR bool, err error) {
 		"CIRCLE_SHA1",
 		"DRONE_COMMIT",
 		"CI_COMMIT_SHA", // GitLab CI
+		"BITBUCKET_COMMIT",
 	})
 	if sha == "" {
 		return nil, false, errors.New("cannot get commit SHA from environment variable. Set CI_COMMIT?")
@@ -84,6 +86,9 @@ func GetBuildInfo() (prInfo *BuildInfo, isPR bool, err error) {
 		"TRAVIS_PULL_REQUEST_BRANCH",
 		"CIRCLE_BRANCH",
 		"DRONE_COMMIT_BRANCH",
+		// present only if PR pipeline
+		"BITBUCKET_PR_DESTINATION_BRANCH",
+		"BITBUCKET_BRANCH",
 	})
 
 	pr := getPullRequestNum()
@@ -134,6 +139,7 @@ func getPullRequestNum() int {
 		"DRONE_PULL_REQUEST",
 		// GitLab CI MergeTrains
 		"CI_MERGE_REQUEST_IID",
+		"BITBUCKET_PR_ID",
 	}
 	// regexp.MustCompile() in func intentionally because this func is called
 	// once for one run.